+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.40.0] - 2020-02-09
 
+### Added
+- Validation: added scroll to invalid DropZone and FileDrop fields
+
 ### Changed
 - README: Updated browsers list in README.md
+
 
 ## [0.39.0] - 2020-01-26